lok3 #4 (smeyer)

+++ Modules/Course/classes/Objectives/class.ilLOSettings.php
@@ -9,6 +9,21 @@
  */
 class ilLOSettings
 {
+	// new settings 5.1
+	const QST_PASSED_FLAG = 1;
+	const QST_PASSED_HIDE = 2;
+	
+	const TYPE_INITIAL_PLACEMENT_ALL = 1;
+	const TYPE_INITIAL_PLACEMENT_SELECTED = 2;
+	const TYPE_INITIAL_QUALIFYING_ALL = 3;
+	const TYPE_INITIAL_QUALIFYING_SELECTED = 4;
+	const TYPE_INITIAL_NONE = 5;
+	
+	const TYPE_QUALIFYING_ALL = 1;
+	const TYPE_QUALIFYING_SELECTED = 2;
+	
+	// end new settings
+	
 	const TYPE_TEST_INITIAL = 1;
 	const TYPE_TEST_QUALIFIED = 2;
 	
@@ -24,12 +39,20 @@ class ilLOSettings
 	
 	private static $instances = array();
 	
+	
+	// settings 5.1
+	private $it_type = self::TYPE_INITIAL_PLACEMENT_ALL;
+	private $qt_type = self::TYPE_QUALIFYING_ALL;
+	
+	private $it_start = FALSE;
+	private $qt_start = FALSE;
+	
+	// end settings 5.1
+	
 	private $container_id = 0;
 	private $type = 0;
 	private $initial_test = 0;
 	private $qualified_test = 0;
-	private $qt_visible_all = true;
-	private $qt_visible_lo = false;
 	private $reset_results = true;
 
 
@@ -58,6 +81,77 @@ class ilLOSettings
 			return self::$instances[$a_obj_id];
 		}
 		return self::$instances[$a_obj_id] = new ilLOSettings($a_obj_id);
+	}
+	
+	/**
+	 * Set Initial test type
+	 * @param type $a_type
+	 */
+	public function setInitialTestType($a_type)
+	{
+		$this->it_type = $a_type;
+	}
+	
+	/**
+	 * Get initial test type
+	 * @return type
+	 */
+	public function getInitialTestType()
+	{
+		return $this->it_type;
+	}
+	
+	/**
+	 * Get qualifying test type
+	 */
+	public function getQualifyingTestType()
+	{
+		return $this->qt_type;
+	}
+	
+	/**
+	 * Set qualifying test type
+	 * @param type $a_type
+	 */
+	public function setQualifyingTestType($a_type)
+	{
+		$this->qt_type = $a_type;
+	}
+	
+	/**
+	 * 
+	 * @param type $a_type
+	 */
+	public function setInitialTestAsStart($a_type)
+	{
+		$this->it_start = $a_type;
+	}
+	
+	/**
+	 * Get initial test start
+	 * @return type
+	 */
+	public function isInitialTestStart()
+	{
+		return $this->it_start;
+	}
+	
+	/**
+	 * Set qt as start object
+	 * @param type $a_type
+	 */
+	public function setQualifyingTestAsStart($a_type)
+	{
+		$this->qt_start = $a_type;
+	}
+	
+	/**
+	 * Is qt start object
+	 * @return type
+	 */
+	public function isQualifyingTestStart()
+	{
+		return $this->qt_start;
 	}
 	
 	/**
@@ -111,6 +205,14 @@ class ilLOSettings
 		
 		$new_settings->create();
 	}
+	
+	/**
+	 * Check if start objects are enabled
+	 */
+	public function worksWithStartObjects()
+	{
+		return $this->isInitialTestStart() or $this->isQualifyingTestStart();
+	}
 
 
 	/**
@@ -118,10 +220,7 @@ class ilLOSettings
 	 */
 	public function worksWithInitialTest()
 	{
-		return 
-				($this->getType() == self::LOC_INITIAL_ALL) or
-				($this->getType() == self::LOC_INITIAL_SEL)
-		;
+		return $this->getInitialTestType() != self::TYPE_INITIAL_NONE;
 	}
 	
 	/**
@@ -130,11 +229,10 @@ class ilLOSettings
 	 */
 	public function isGeneralQualifiedTestVisible()
 	{
-		return $this->qt_visible_all;
+		return $this->getQualifyingTestType() == self::TYPE_QUALIFYING_ALL;
 	}
 
 	/**
-	 * Check if qualified test for all objectives is visible
 	 * @return type
 	 */
 	public function setGeneralQualifiedTestVisibility($a_stat)
@@ -145,7 +243,7 @@ class ilLOSettings
 	
 	public function isQualifiedTestPerObjectiveVisible()
 	{
-		return $this->qt_visible_lo;
+		return $this->getQualifyingTestType() == self::TYPE_QUALIFYING_SELECTED;
 	}
 	
 	public function setQualifiedTestPerObjectiveVisibility($a_stat)
@@ -177,7 +275,12 @@ class ilLOSettings
 		return $this->type;
 	}
 	
-	
+	/**
+	 * 
+	 * @param type $a_type
+	 * @return type
+	 * @todo refactor
+	 */
 	public function getTestByType($a_type)
 	{
 		switch($a_type)
@@ -193,6 +296,7 @@ class ilLOSettings
 	/**
 	 * Get assigned tests
 	 * @return type
+	 * @todo refactor
 	 */
 	public function getTests()
 	{
@@ -212,6 +316,7 @@ class ilLOSettings
 	 * Check if test is of type random test
 	 * @param type $a_type
 	 * @return type
+	 * @todo refactor
 	 */
 	public function isRandomTestType($a_type)
 	{
@@ -223,32 +328,56 @@ class ilLOSettings
 	/**
 	 * set initial test id
 	 * @param type $a_id
+	 * @todo refactor
 	 */
 	public function setInitialTest($a_id)
 	{
 		$this->initial_test = $a_id;
 	}
-	
+
+	/**
+	 * get initial test
+	 * @return type
+	 * @todo refactor
+	 */
 	public function getInitialTest()
 	{
 		return $this->initial_test;
 	}
-	
+
+	/**
+	 * set qualified test
+	 * @param type $a_id
+	 * @todo refactor
+	 */
 	public function setQualifiedTest($a_id)
 	{
 		$this->qualified_test = $a_id;
 	}
 	
+	/**
+	 * get qualified test
+	 * @return type
+	 * @todo refactor
+	 */
 	public function getQualifiedTest()
 	{
 		return $this->qualified_test;
 	}
-	
+
+	/**
+	 * reset results
+	 * @param type $a_status
+	 */
 	public function resetResults($a_status)
 	{
 		$this->reset_results = $a_status;
 	}
 	
+	/**
+	 * check if reset result is enabled
+	 * @return type
+	 */
 	public function isResetResultsEnabled()
 	{
 		return (bool) $this->reset_results;
@@ -262,13 +391,14 @@ class ilLOSettings
 		global $ilDB;
 		
 		$query = 'INSERT INTO loc_settings '.
-				'(obj_id, type,itest,qtest,qt_vis_all,qt_vis_obj,reset_results) VALUES ( '.
+				'(obj_id, it_type,itest,qtest,it_start,qt_type,qt_start,reset_results) VALUES ( '.
 				$ilDB->quote($this->getObjId(),'integer').', '.
-				$ilDB->quote($this->getType(),'integer').', '.
+				$ilDB->quote($this->getInitialTestType(),'integer').', '.
 				$ilDB->quote($this->getInitialTest(),'integer').', '.
 				$ilDB->quote($this->getQualifiedTest(),'integer').', '.
-				$ilDB->quote($this->isGeneralQualifiedTestVisible(),'integer').', '.
-				$ilDB->quote($this->isQualifiedTestPerObjectiveVisible(),'integer').', '.
+				$ilDB->quote($this->isInitialTestStart(),'integer').', '.
+				$ilDB->quote($this->getQualifyingTestType(),'integer').', '.
+				$ilDB->quote($this->isQualifyingTestStart(),'integer').', '.
 				$ilDB->quote($this->isResetResultsEnabled(),'integer').' '.
 				') ';
 		$ilDB->manipulate($query);
@@ -289,11 +419,12 @@ class ilLOSettings
 		}
 		
 		$query = 'UPDATE loc_settings '.' '.
-				'SET type = '.$ilDB->quote($this->getType(),'integer').', '.
+				'SET it_type = '.$ilDB->quote($this->getInitialTestType(),'integer').', '.
 				'itest = '.$ilDB->quote($this->getInitialTest(),'integer').', '.
 				'qtest = '.$ilDB->quote($this->getQualifiedTest(),'integer').', '.
-				'qt_vis_all = '.$ilDB->quote($this->isGeneralQualifiedTestVisible(),'integer').', '.
-				'qt_vis_obj = '.$ilDB->quote($this->isQualifiedTestPerObjectiveVisible(),'integer').', '.
+				'it_start = '.$ilDB->quote($this->isInitialTestStart(),'integer').', '.
+				'qt_type = '.$ilDB->quote($this->getQualifyingTestType(),'integer').', '.
+				'qt_start = '.$ilDB->quote($this->isQualifyingTestStart(),'integer').', '.
 				'reset_results = '.$ilDB->quote($this->isResetResultsEnabled(),'integer').' '.
 				'WHERE obj_id = '.$ilDB->quote($this->getObjId(),'integer');
 				
@@ -308,43 +439,73 @@ class ilLOSettings
 	 */
 	public function updateStartObjects(ilContainerStartObjects $start)
 	{
-		switch($this->getType())
+		if($this->getInitialTestType() != self::TYPE_INITIAL_NONE)
 		{
-			case self::LOC_INITIAL_ALL:
-				if($start->exists($this->getQualifiedTest()))
+			if($start->exists($this->getQualifiedTest()))
+			{
+				$start->deleteItem($this->getQualifiedTest());
+			}
+		}
+		
+		switch($this->getInitialTestType())
+		{
+			case self::TYPE_INITIAL_PLACEMENT_ALL:
+			case self::TYPE_INITIAL_QUALIFYING_ALL:
+				
+				if($this->isInitialTestStart())
 				{
-					$start->deleteItem($this->getQualifiedTest());
+					if(!$start->exists($this->getInitialTest()))
+					{
+						$start->add($this->getInitialTest());
+					}
 				}
-				if(!$start->exists($this->getInitialTest()))
+				else
 				{
-					$start->add($this->getInitialTest());
+					if($start->exists($this->getInitialTest()))
+					{
+						$start->deleteItem($this->getInitialTest());
+					}
 				}
 				break;
 				
-			case self::LOC_INITIAL_SEL:
-			case self::LOC_PRACTISE:
-				if($start->exists($this->getQualifiedTest()))
-				{
-					$start->deleteItem($this->getQualifiedTest());
-				}
+			case self::TYPE_INITIAL_NONE:
+				
 				if($start->exists($this->getInitialTest()))
 				{
 					$start->deleteItem($this->getInitialTest());
 				}
 				break;
 				
-			case self::LOC_QUALIFIED:
-				if(!$start->exists($this->getQualifiedTest()))
-				{
-					$start->add($this->getQualifiedTest());
-				}
+			default:
+
 				if($start->exists($this->getInitialTest()))
 				{
 					$start->deleteItem($this->getInitialTest());
 				}
 				break;
 		}
-		return true;
+		
+		switch($this->getQualifyingTestType())
+		{
+			case self::TYPE_QUALIFYING_ALL:
+				
+				if($this->isQualifyingTestStart())
+				{
+					if(!$start->exists($this->getQualifiedTest()))
+					{
+						$start->add($this->getQualifiedTest());
+					}
+				}
+				break;
+				
+			default:
+				if($start->exists($this->getQualifiedTest()))
+				{
+					$start->deleteItem($this->getQualifiedTest());
+				}
+				break;
+		}
+		return TRUE;
 	}
 	
 	
@@ -361,14 +522,19 @@ class ilLOSettings
 		while($row = $res->fetchRow(DB_FETCHMODE_OBJECT))
 		{
 			$this->entry_exists = true;
-			$this->setType($row->type);
+			
+			$this->setInitialTestType($row->it_type);
+			$this->setInitialTestAsStart((bool) $row->it_start);
+			$this->setQualifyingTestType($row->qt_type);
+			$this->setQualifyingTestAsStart($row->qt_start);
+			
+			#$this->setType($row->type);
 			$this->setInitialTest($row->itest);
 			$this->setQualifiedTest($row->qtest);
 			#$this->setGeneralQualifiedTestVisibility($row->qt_vis_all);
-			$this->setQualifiedTestPerObjectiveVisibility($row->qt_vis_obj);
+			#$this->setQualifiedTestPerObjectiveVisibility($row->qt_vis_obj);
 			$this->resetResults($row->reset_results);
 		}
-		
 		
 		if($GLOBALS['tree']->isDeleted($this->getInitialTest()))
 		{
